✨ Feature: 하루 건강 데이터 전송 및 건강 레포트 조회 구현

+++ WAYVI/WAYVI Watch App/Models/Auth/Response/UserResponse.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct LoginResponse: Codable {
     let status: String
-    let userId: Int64
+    let userId: Int
 }
